@@ -1,75 +1,87 @@
 {
-    "name": "@acaprojects/ngx-widgets",
-    "version": "0.14.21",
-    "license": "MIT",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e",
-        "packagr": "ng-packagr -p ng-package.json"
-    },
-    "dependencies": {
-        "core-js": "^2.4.1",
-        "cropperjs": "^1.3.2",
-        "hammerjs": "^2.0.8",
-        "moment": "^2.21.0",
-        "rxjs": "^5.5.7",
-        "zone.js": "^0.8.20"
-    },
-    "peerDependancies": {
-        "@angular/animations": "^5.2.9",
-        "@angular/common": "^5.2.9",
-        "@angular/compiler": "^5.2.9",
-        "@angular/core": "^5.2.9",
-        "@angular/forms": "^5.2.9",
-        "@angular/http": "^5.2.9",
-        "@angular/platform-browser": "^5.2.9",
-        "@angular/platform-browser-dynamic": "^5.2.9",
-        "@angular/router": "^5.2.9"
-    },
-    "devDependencies": {
-        "@angular/animations": "^5.2.9",
-        "@angular/common": "^5.2.9",
-        "@angular/compiler": "^5.2.9",
-        "@angular/core": "^5.2.9",
-        "@angular/forms": "^5.2.9",
-        "@angular/http": "^5.2.9",
-        "@angular/platform-browser": "^5.2.9",
-        "@angular/platform-browser-dynamic": "^5.2.9",
-        "@angular/router": "^5.2.9",
-        "@angular/cli": "^1.7.3",
-        "@angular/compiler-cli": "^5.2.9",
-        "@angular/language-service": "^5.2.9",
-        "@types/jasmine": "^2.8.6",
-        "@types/jasminewd2": "~2.0.2",
-        "@types/node": "^6.0.102",
-        "codelyzer": "^4.2.1",
-        "electron-packager": "^10.1.2",
-        "gulp": "^3.9.1",
-        "gulp-bump": "^2.9.0",
-        "gulp-install": "^1.1.0",
-        "gulp-json-modify": "^1.0.2",
-        "gulp-typescript": "^3.2.4",
-        "gulp-util": "^3.0.8",
-        "highlight.js": "^9.12.0",
-        "jasmine-core": "~2.6.2",
-        "jasmine-spec-reporter": "~4.1.0",
-        "karma": "~1.7.0",
-        "karma-chrome-launcher": "~2.1.1",
-        "karma-cli": "~1.0.1",
-        "karma-coverage-istanbul-reporter": "^1.4.2",
-        "karma-jasmine": "^1.1.1",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "merge": "^1.2.0",
-        "ng-packagr": "^2.2.0",
-        "protractor": "~5.1.2",
-        "require-dir": "^0.3.2",
-        "run-sequence": "^2.2.1",
-        "ts-node": "~3.2.0",
-        "tslint": "~5.7.0",
-        "typescript": "^2.7.2"
-    }
+  "name": "@acaprojects/ngx-widgets",
+  "version": "0.15.2",
+  "license": "MIT",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "packagr": "ng-packagr -p ng-package.json"
+  },
+  "dependencies": {
+    "cropperjs": "^1.3.6",
+    "hammerjs": "^2.0.8",
+    "moment": "^2.22.1"
+  },
+  "peerDependancies": {
+    "@angular/animations": "^6.0.2",
+    "@angular/cli": "^6.0.3",
+    "@angular/common": "^6.0.2",
+    "@angular/compiler": "^6.0.2",
+    "@angular/compiler-cli": "^6.0.2",
+    "@angular/core": "^6.0.2",
+    "@angular/forms": "^6.0.2",
+    "@angular/http": "^6.0.2",
+    "@angular/language-service": "^6.0.2",
+    "@angular/platform-browser": "^6.0.2",
+    "@angular/platform-browser-dynamic": "^6.0.2",
+    "@angular/router": "^6.0.2",
+    "rxjs": "^6.1.0",
+    "ts-node": "~3.2.0",
+    "tslint": "~5.7.0",
+    "typescript": "~2.7.2",
+    "zone.js": "^0.8.26",
+    "cropperjs": "^1.3.4",
+    "hammerjs": "^2.0.8",
+    "moment": "^2.22.0"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.6.3",
+    "@angular/animations": "^6.0.2",
+    "@angular/cli": "^6.0.3",
+    "@angular/common": "^6.0.2",
+    "@angular/compiler": "^6.0.2",
+    "@angular/compiler-cli": "^6.0.2",
+    "@angular/core": "^6.0.2",
+    "@angular/forms": "^6.0.2",
+    "@angular/http": "^6.0.2",
+    "@angular/language-service": "^6.0.2",
+    "@angular/platform-browser": "^6.0.2",
+    "@angular/platform-browser-dynamic": "^6.0.2",
+    "@angular/router": "^6.0.2",
+    "@types/jasmine": "^2.8.7",
+    "@types/jasminewd2": "~2.0.2",
+    "@types/node": "^6.0.111",
+    "codelyzer": "^4.3.0",
+    "core-js": "^2.5.6",
+    "electron-packager": "^10.1.2",
+    "gulp": "^3.9.1",
+    "gulp-bump": "^2.9.0",
+    "gulp-install": "^1.1.0",
+    "gulp-json-modify": "^1.0.2",
+    "gulp-typescript": "^3.2.4",
+    "gulp-util": "^3.0.8",
+    "highlight.js": "^9.12.0",
+    "jasmine-core": "~2.6.2",
+    "jasmine-spec-reporter": "~4.1.0",
+    "karma": "~1.7.0",
+    "karma-chrome-launcher": "~2.1.1",
+    "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^1.4.3",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "merge": "^1.2.0",
+    "ng-packagr": "^2.4.5",
+    "protractor": "~5.1.2",
+    "require-dir": "^0.3.2",
+    "run-sequence": "^2.2.1",
+    "rxjs": "^6.1.0",
+    "ts-node": "~3.2.0",
+    "tslint": "~5.7.0",
+    "typescript": "~2.7.2",
+    "zone.js": "^0.8.26"
+  }
 }
