@@ -10,7 +10,7 @@ export class CheckboxComponent {
     @Input() name = '';
     @Input() model = false;
     @Input() label = '';
-    @Input() side = 'left';
+    @Input() side = 'right';
     @Output() modelChange: any = new EventEmitter();
 
     public toggle() {
