@@ -9,7 +9,9 @@ import { TooltipComponent } from '../components/overlays/tooltip/tooltip.compone
 
 import { WIDGETS } from '../settings';
 
-@Injectable()
+@Injectable({
+    providedIn: 'root'
+})
 export class OverlayService {
     private cmp_reg: any = {};
     private cmp_list: any = {};
@@ -307,8 +309,7 @@ export class OverlayService {
             this.containers.root = cmp.instance;
             this.containers.root.ng = cmp;
             this.containers.root.service = this;
-            this.add('root', 'notifications', NotificationComponent, {})
-                .then(() => null, () => null);
+            this.add('root', 'notifications', NotificationComponent, {}).then(() => null, () => null);
             return;
         } else if (!this._view && this.default_vc) {
             this._view = this.default_vc;
