@@ -28,7 +28,7 @@ import { ClickResponderComponent } from './components/click-responder/click-resp
 import { CalendarComponent } from './components/form-controls/calendar/calendar.component';
 import { CheckboxComponent } from './components/form-controls/checkbox/checkbox.component';
 import { CustomDropdownComponent } from './components/form-controls/custom-dropdown/dropdown.component';
-import { DataInputComponent } from './components/form-controls/data-input/data-input.component';
+import { InputFieldComponent } from './components/form-controls/input-field/input-field.component';
 import { DropdownComponent } from './components/form-controls/dropdown/dropdown.component';
 import { RadioButtonComponent } from './components/form-controls/radioset/radio-button/radio-button.component';
 import { RadiosetComponent } from './components/form-controls/radioset/radioset.component';
@@ -68,6 +68,8 @@ import { MapInputDirective } from './directives/map-input.directive';
 import { TooltipDirective } from './directives/overlays/tooltip.directive';
 import { ModalDirective } from './directives/overlays/modal.directive';
 import { NotifyDirective } from './directives/overlays/notify.directive';
+import { TapDownDirective } from './directives/tapdown.directive';
+import { TapUpDirective } from './directives/tapup.directive';
     // Pipes
 import { KeysPipe } from './pipes/keys.pipe';
 import { SafeStylePipe } from './pipes/safe-style.pipe';
@@ -101,7 +103,7 @@ export class WidgetsHammerConfig extends HammerGestureConfig {
         CalendarComponent,
         CheckboxComponent,
         CustomDropdownComponent,
-        DataInputComponent,
+        InputFieldComponent,
         DropdownComponent,
         RadioButtonComponent,
         RadiosetComponent,
@@ -128,6 +130,8 @@ export class WidgetsHammerConfig extends HammerGestureConfig {
         TooltipDirective,
         ModalDirective,
         NotifyDirective,
+        TapDownDirective,
+        TapUpDirective,
             // Declare Pipes
         // ...PIPES,
         KeysPipe,
@@ -163,7 +167,7 @@ export class WidgetsHammerConfig extends HammerGestureConfig {
         CalendarComponent,
         CheckboxComponent,
         CustomDropdownComponent,
-        DataInputComponent,
+        InputFieldComponent,
         DropdownComponent,
         RadioButtonComponent,
         RadiosetComponent,
@@ -190,6 +194,9 @@ export class WidgetsHammerConfig extends HammerGestureConfig {
         TooltipDirective,
         ModalDirective,
         NotifyDirective,
+        TapDownDirective,
+        TapUpDirective,
+        ClickResponderComponent,
             // Export Pipes
         // ...PIPES,
         KeysPipe,
@@ -232,7 +239,7 @@ export class WidgetsHammerConfig extends HammerGestureConfig {
 })
 export class WidgetsModule {
     private static init = false;
-    private build = '2018-04-03.v1';
+    private build = '2018-05-23.v1';
 
     constructor() {
         if (!WidgetsModule.init) {
@@ -245,11 +252,6 @@ export class WidgetsModule {
         return {
             ngModule: WidgetsModule,
             providers: [
-                // ...SERVICES
-                Animate,
-                MapService,
-                OverlayService,
-                DropService,
                 {
                     provide: HAMMER_GESTURE_CONFIG,
                     useClass: WidgetsHammerConfig
